Extract text resources and refactor code

+++ src/main/java/com/devserbyn/suncheck/service/impl/UserServiceImpl.java
@@ -1,7 +1,9 @@
 package com.devserbyn.suncheck.service.impl;
 
+import com.devserbyn.suncheck.constant.INTEGER_CONSTANT;
 import com.devserbyn.suncheck.model.User;
 import com.devserbyn.suncheck.model.UserConfig;
+import com.devserbyn.suncheck.model.UserSetupStepsBO;
 import com.devserbyn.suncheck.repository.UserConfigRepository;
 import com.devserbyn.suncheck.service.UserRepositoryService;
 import com.devserbyn.suncheck.service.UserService;
@@ -9,10 +11,13 @@ import com.devserbyn.suncheck.service.UserService;
 import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
-import java.util.Optional;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+
+import static com.devserbyn.suncheck.constant.INTEGER_CONSTANT.REMOVEME_EXPIRE_SEC;
 
 @Service
 @Slf4j
@@ -30,7 +35,7 @@ public class UserServiceImpl implements UserService {
         try {
             userRepoService.save(newUser).orElseThrow(RuntimeException::new);
         } catch (Exception e) {
-            System.out.println("User registration failed");
+            log.error("User registration failed", e);
         }
     }
 
@@ -39,10 +44,14 @@ public class UserServiceImpl implements UserService {
         Long userChatId = update.getMessage().getChatId();
         UserConfig userConfig = userConfigRepository.findAll().stream()
                 .filter(c -> c.getUser().getChatId() == userChatId)
-                .findFirst().orElseThrow(() -> new RuntimeException("User config wasn't found"));
+                .findFirst().orElse(null);
         try {
-            userConfigRepository.delete(userConfig);
-            userRepoService.deleteByChatId(userChatId);
+            if (userConfig != null) {
+                userConfigRepository.delete(userConfig);
+            }
+            if (this.checkIfUserRegistered(userChatId)) {
+                userRepoService.deleteByChatId(userChatId);
+            }
         } catch (Exception e) {
             userConfigRepository.save(userConfig);
             log.error("User removing failed", e);
